Pruebas API RestClient_Unity
Se actualiza el juego con pruebas exitosas de envió de datos hacia una RealTimeDB alojada en Firebase

+++ Assets/Scripts/CollecObjetos.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class CollecObjetos : MonoBehaviour
 {
    [SerializeField] private GameObject efecto;
-   [SerializeField] private float  cantidadPuntos;
+   [SerializeField] private int  cantidadPuntos;
    [SerializeField] private Puntaje puntaje;
